Ajout d'un mixin pour gérer la prévisualisation dans le module tutov2
